Fixed bug in IE when using relative URIs
When using relative redirect URIs, in IE the HTMLAnchorElement created using said URI has as hostname <host>:<port> even when the default port is used by the Website. Causes the popup not to close because the URIs don't match. Now the hostname used only has the port appended if it's not 80 and 443